requirements.txt
pydantic>=2.0.0
click>=8.0.0
rich>=13.0.0
pytest>=7.0.0
pyyaml>=6.0.0
requests>=2.0.0
notebook>=6.0.0
python-dotenv>=0.19.0
pandas
plotly
pyyaml